+++ scripts/test-zsh-integration.zsh
@@ -276,6 +276,9 @@ typeset -g last_notice_message='unchanged notice'
 _aicoach_handle_line $'PRIVACY_RECEIPT\t'$AICOACH_SESSION_ID$'\treq-privacy\tchat\tsucceeded\t128\t2\t1\ttrue\t20'
 assert_eq "$BUFFER" 'echo unchanged'
 assert_eq "$last_notice_message" 'unchanged notice'
+_aicoach_handle_line $'AIRLOCK\t'$AICOACH_SESSION_ID$'\t\tfalse\t1'
+assert_eq "$BUFFER" 'echo unchanged'
+assert_eq "$last_notice_message" 'unchanged notice'
 
 (( test_failed == 0 )) && print 'zsh integration tests: ok'
 exit $test_failed